refactor : Cars - maxMove 로직 재구성
