+++ src/main/java/racingcar/game/view/InputView.java
@@ -1,7 +1,7 @@
 package racingcar.game.view;
 
 import camp.nextstep.edu.missionutils.Console;
-import racingcar.domain.car.Cars;
+import racingcar.domain.cars.Cars;
 import racingcar.domain.race.RaceCount;
 
 public class InputView {
@@ -13,7 +13,7 @@ public class InputView {
 
     public Cars getCars() {
         String readLine = Console.readLine();
-        
+
         return Cars.fromNames(readLine.split(","));
     }
 }
