+++ src/test/java/racingcar/domain/car/CarsTest.java
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import racingcar.domain.car.move.NumberCarMove;
+import racingcar.domain.cars.Cars;
 
 public class CarsTest {
     Car pobi;
@@ -31,14 +32,14 @@ public class CarsTest {
         cars.move(NumberCarMove.forward());
 
         Assertions.assertArrayEquals(cars
-                .toMaxMoved()
+                .filterMaxMoved()
                 .getNames()
                 .toArray(), new String[]{"pobi", "jun"});
 
         pobi.move(NumberCarMove.forward());
 
         Assertions.assertArrayEquals(cars
-                .toMaxMoved()
+                .filterMaxMoved()
                 .getNames()
                 .toArray(), new String[]{"pobi"});
     }
