+++ src/test/java/racingcar/domain/race/RaceTest.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import racingcar.domain.car.Car;
-import racingcar.domain.car.Cars;
+import racingcar.domain.cars.Cars;
 import racingcar.domain.car.move.NumberCarMove;
 
 public class RaceTest {
